requirements.txt
torch
transformers
timm
streamlit
opencv-python
Pillow
numpy
